% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxplot_data.R
\name{Boxplot_data}
\alias{Boxplot_data}
\title{Creating Boxplot for a dataset}
\usage{
Boxplot_data(data)
}
\arguments{
\item{data}{Proteomics expression dataset (original or normalized dataset)}
}
\value{
Interactive box and whiskers plot
}
\description{
The box and whiskers plot displays the distribution of a continuous variable. 
It visualises five summary statistics (the median, two hinges and two whiskers), and all "outlying" points individually. 
The `ggplot2` package is used here for creating the boxplot.
}
\details{
This can also be used for comparing the original dataset with the normalized dataset.
}
\examples{
Boxplot_data(yeast_data)
Boxplot_data(knn_rlr_yeast_data)
}
\seealso{
`geom_boxplot()`
}
