% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corrplot_data.R
\name{Corrplot_data}
\alias{Corrplot_data}
\title{Creating Correlation matrix plot for a dataset}
\usage{
Corrplot_data(data)
}
\arguments{
\item{data}{Proteomics expression dataset (original or normalized dataset) along with the protein information}
}
\value{
Interactive corrleation matrix plot
}
\description{
A graphical display of a correlation matrix.
}
\details{
This can also be used for comparing the original dataset with the normalized dataset.
}
\examples{
Corrplot_data(yeast_data)
Corrplot_data(knn_rlr_yeast_data)
}
