% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Densityplot_data.R
\name{Densityplot_data}
\alias{Densityplot_data}
\title{Creating Density plot for a dataset}
\usage{
Densityplot_data(data)
}
\arguments{
\item{data}{Proteomics expression dataset (original or normalized dataset) along with the protein information}
}
\value{
Interactive column-wise density plot
}
\description{
Computes and draws kernel density estimate, which is a smoothed version of the histogram. 
This is a useful alternative to the histogram for continuous data that comes from an underlying smooth distribution. 
The `ggplot2` package is used here for creating the boxplot.
}
\details{
This can also be used for comparing the original dataset with the normalized dataset.
}
\examples{
Densityplot_data(yeast_data)
Densityplot_data(knn_rlr_yeast_data)
}
\seealso{
`geom_density()`
}
