% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot_data.R
\name{MDSplot_data}
\alias{MDSplot_data}
\title{Creating MDS plot for a dataset}
\usage{
MDSplot_data(data)
}
\arguments{
\item{data}{Normalized and imputed Proteomics expression dataset along with the protein information}
}
\value{
MDS plot
}
\description{
Multi-dimensional scaling (MDS) plots showing a 2-dimensional projection of distances between the dataset samples.
}
\examples{
MDSplot_data(knn_rlr_yeast_data)
}
\seealso{
`mdsPlot`
}
