% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot_data.R
\name{QQplot_data}
\alias{QQplot_data}
\title{Creating QQ-Plot for a dataset}
\usage{
QQplot_data(data)
}
\arguments{
\item{data}{Proteomics expression dataset (original or normalized dataset)}
}
\value{
Interactive column-wise QQ-plot
}
\description{
A Q–Q plot (quantile-quantile plot) is a plot of the quantiles of two distributions against each other, or a plot based on estimates of the quantiles. 
The normality of the data can be understand by this plot.
}
\details{
This can be used for comparing the original dataset with the
normalized dataset.
}
\examples{
qqplot <- QQplot_data(rlr_knn_yeast_data)
}
