% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_groups.R
\docType{data}
\name{yeast_groups}
\alias{yeast_groups}
\title{Group information of Yeast lysate - UPS1 benchmark dataset}
\format{
A data frame with 6 rows and 2 variables:
\describe{
  \item{Samples}{Technical replicate details - Two kinds of technical replicates
  (A1, A2, A3) and (B1, B2, B3)}
  \item{Groups}{Condition details - Two conditions (GrA) and (GrB)}
}
}
\source{
\doi{10.1016/j.dib.2015.11.063}
}
\usage{
yeast_groups
}
\description{
The standard benchmark yeast lysate - UPS1 dataset contains the variables of
6 columns, 2 conditions and 3 technical replicates.
}
\details{
The above information is required for further analysis of the dataset.
}
\keyword{datasets}
