% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_top_table.R
\docType{data}
\name{yeast_top_table}
\alias{yeast_top_table}
\title{Top table for Yeast Lysate - UPS1 dataset}
\format{
A data frame with 835 rows and 6 variables along with 'Fasta headers':
\describe{
  \item{logFC}{Log fold change values}
  \item{AveExpr}{Average intensity values}
  \item{t}{t-statistic values}
  \item{P.Value}{P-values}
  \item{adj.P.Val}{Adjusted P-values}
  \item{B}{B-statistic values}
}
}
\usage{
yeast_top_table
}
\description{
Top table can be used for identifying the differential abundance analysis of proteins in the dataset.
}
\keyword{datasets}
